requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.40.0
idna==2.8
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
sqlparse==0.3.0
urllib3==1.25.6
iamport-rest-client==0.7.0
pillow==6.2.0
djangorestframework==3.10.3
django-robots==3.1.0
django-configurations==2.1
django-debug-toolbar==2.0
django-axes==5.0.18
django-pipeline==1.6.14
django-extensions==2.2.3
django-storages==1.7.2
psycopg2==2.8.4
boto3==1.9.253
gunicorn==19.9.0
sentry-sdk==0.13.0
django-admin-honeypot==1.1.0
